refact: 戻るボタンのテストを別ファイルに切り出した
